Auto-generate properties for ..._html, ..._label_key and ..._label_value, deprecate UnitModelMixin
The use of __getattr__ in UnitModelMixin was error prone. Now properties
are generated automatically for accessing special output of fields.

While this is still not perfectly safe (an attribute with the same name
could exist in the class already), it's better than blindly looking for
suffices like "_html" like it used to be before.